+++ src/Controller/CharacterController.php
@@ -10,7 +10,7 @@ use App\Entity\Character;
 final class CharacterController extends AbstractController
 {
     // DISPLAY
-    #[Route('/characters/', name: 'app_character_display')]
+    #[Route('/characters/', name: 'app_character_display', methods: ['GET'])]
     public function display(): JsonResponse
     {
         $character = new Character();