+++ Changelog.md
@@ -1,3 +1,10 @@
 ChangeLog
 v0.1
  - Core file test
+V0.2
+- Ajout Entity Character
+- Ajout .htaccess
+- Ajout VarDumper (dev)
+- Ajout CharacterController
+- Ajout tests
+- Ajout des methods à la route